Rename validate_results file
