+++ src/util/mod.rs
@@ -3,8 +3,8 @@
 mod convergence_results;
 mod paraview;
 mod reference_data;
-mod validate_displacements;
+mod validate_results;
 pub use crate::util::convergence_results::*;
 pub use crate::util::paraview::*;
 pub use crate::util::reference_data::*;
-pub use crate::util::validate_displacements::*;
+pub use crate::util::validate_results::*;
