+++ src/util/validate_results.rs
